Move handler combinators out of Frontend.Util
